Allow plugins to modify its own webpack config
This adds a hook for plugins to modify the webpack config used to build
its own bundle.  The semantics for this are the same for pre 3.0
plugins; you provide a module containing a function that mutates the
predefined webpack config arbitrarily.  Now plugin authors must
explicitly define the module containing the webpack helper function.