Fix python package install on readthedocs
By default, readthedocs runs `setup.py install` which is subtly
different than `pip install .`.  A recent change in click (I think)
broke the package install step on rtd's build machine.  This uses the
new readthedocs config file to force installation with pip.  It seems to
be the only way to configure this option.

https://docs.readthedocs.io/en/latest/yaml-config.html