Add setuptools-git as an install time requirement for plugins
This ensures that web_client assets are always installed with the python
package.  Alternatively, we could recommend a package_data kwarg to
setup, but IMO that option is confusing.  With this change, using
`include_package_data` with `setuptools-git` (or an explicit manifest)
is the recommended method for including web client assets for external
plugins.