Use the real path for web builds.
In a virtual Python 2.7 environment, with Girder installed in
non-editable mode, the web_client path contains <env
root>/local/lib/python2.7/site-packages/girder/web_client, but local/lib
is a symlink to lib.  When npm installs packages, it fails to put the
packages in the expect spot, so, for instance, underscore ends up in
.../web_client/node_modules/@girder/core/node_modules/underscore rather
than an .../web_client/node_modules/underscore.  This causes the web
client build to fail, as webpack fails to locate some modules.